+++ app/services/mask_generation.py
@@ -1,9 +1,6 @@
 import json
 
-from fastapi import Depends
-from sqlalchemy.orm import Session
-
-from app.database import get_context_session, get_session
+from app.database import get_context_session
 from app.database.images import Images
 from app.database.masks import Masks
 from app.database.contours import Contours
@@ -11,9 +8,6 @@ from app.database.labels import Labels
 import numpy as np
 import cv2 as cv
 
-from app.routes.contours import add_contours
-from app.routes.masks import create_mask
-from app.schemas.segmentation.segmentations import SegmentationMaskModel
 from app.services.contours import build_depth_first_contour_list
 
 
@@ -38,37 +32,3 @@ def generate_mask(mask_id):
             cv_contour[..., 1] *= image.height
             cv.fillPoly(canvas, [cv_contour.astype(np.int32)], color=[label_id_to_value[contour.label]])
         return canvas
-
-
-async def create_masks_and_add_contours_for_images(image_ids: list[int],
-                                                   mask_responses: list[SegmentationMaskModel],
-                                                   db: Session = Depends(get_session)):
-    """
-    Create masks for a list of image IDs and add contours to them.
-
-    Args:
-        image_ids (list[int]): List of image IDs for which to create masks.
-        mask_responses (list[SegmentationMaskModel]): List of segmentation mask responses containing contours.
-        db (Session): The database session.
-
-    Returns:
-        dict: A dictionary containing the success status, message, and responses for each image.
-    """
-    if len(image_ids) != len(mask_responses):
-        raise ValueError(
-            f"Number of image_ids does not match number of mask_responses."
-        )
-    responses = []
-    for image_id, mask_response in zip(image_ids, mask_responses):
-        mask = db.query(Masks).filter_by(image_id=image_id).first()
-        if not mask:
-            response = await create_mask(image_id, db)
-            mask = db.query(Masks).filter_by(image_id=image_id).first()
-        responses.append(await add_contours(mask.id, mask_response.contours, None, db))
-        mask.generated = True
-        db.commit()
-    return {
-        "success": True,
-        "message": f"Created and added masks for {len(image_ids)} images.",
-        "responses": responses
-    }
